face_recognition_system/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Deep learning and computer vision
facenet-pytorch>=2.5.3
mtcnn>=0.1.1
faiss-cpu>=1.7.4

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data processing and analysis
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
redis>=5.0.0
celery>=5.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Frontend (if building integrated solution)
# Note: Frontend will be separate React app